cheri: tools: perf Use intptr_t instead of long in hashmap interface
Commit c302378bc157 ("libbpf: Hashmap interface update to allow both
long and void* keys/values") introduced a polymorphic hashmap interface
that casts keys and values between `void *` and `long`.

This introduced a hard assumption that `sizeof(long) == sizeof(void *)`.
But it breaks compilation and runtime safety on architectures where
`sizeof(void *) > sizeof(long)`, such as CHERI system.

On CHERI, pointers are 128-bit capabilities while `long` remains 64-bit.
Casting a 128-bit capability pointer to a 64-bit `long` strips the
hardware validity tag and truncates the metadata. Furthermore, the
`_Static_assert` introduced in the previous commit explicitly fails
the build on these architectures.

Fix this by replacing `long` with `intptr_t` throughout the hashmap
interface and related macros. `intptr_t` guarantees safe
pointer-to-interger round-tripping without metadata loss or truncation.
This restores compiler compatibility and runtime safety for CHERI
architectures.

+++ tools/perf/builtin-ftrace.c
@@ -1193,12 +1193,12 @@ out:
 	return (done && !workload_exec_errno) ? 0 : -1;
 }
 
-static size_t profile_hash(long func, void *ctx __maybe_unused)
+static size_t profile_hash(intptr_t func, void *ctx __maybe_unused)
 {
 	return str_hash((char *)func);
 }
 
-static bool profile_equal(long func1, long func2, void *ctx __maybe_unused)
+static bool profile_equal(intptr_t func1, intptr_t func2, void *ctx __maybe_unused)
 {
 	return !strcmp((char *)func1, (char *)func2);
 }
