+++ tools/perf/util/fncache.c
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0-only
 /* Manage a cache of file names' existence */
 #include <pthread.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
@@ -10,12 +11,12 @@
 
 static struct hashmap *fncache;
 
-static size_t fncache__hash(long key, void *ctx __maybe_unused)
+static size_t fncache__hash(intptr_t key, void *ctx __maybe_unused)
 {
 	return str_hash((const char *)key);
 }
 
-static bool fncache__equal(long key1, long key2, void *ctx __maybe_unused)
+static bool fncache__equal(intptr_t key1, intptr_t key2, void *ctx __maybe_unused)
 {
 	return strcmp((const char *)key1, (const char *)key2) == 0;
 }
@@ -36,7 +37,7 @@ static struct hashmap *fncache__get(void)
 
 static bool lookup_fncache(const char *name, bool *res)
 {
-	long val;
+	intptr_t val;
 
 	if (!hashmap__find(fncache__get(), name, &val))
 		return false;
