+++ tools/perf/util/hashmap.c
@@ -128,7 +128,7 @@ static int hashmap_grow(struct hashmap *map)
 }
 
 static bool hashmap_find_entry(const struct hashmap *map,
-			       const long key, size_t hash,
+			       const intptr_t key, size_t hash,
 			       struct hashmap_entry ***pprev,
 			       struct hashmap_entry **entry)
 {
@@ -151,9 +151,9 @@ static bool hashmap_find_entry(const struct hashmap *map,
 	return false;
 }
 
-int hashmap_insert(struct hashmap *map, long key, long value,
+int hashmap_insert(struct hashmap *map, intptr_t key, intptr_t value,
 		   enum hashmap_insert_strategy strategy,
-		   long *old_key, long *old_value)
+		   intptr_t *old_key, intptr_t *old_value)
 {
 	struct hashmap_entry *entry;
 	size_t h;
@@ -203,7 +203,7 @@ int hashmap_insert(struct hashmap *map, long key, long value,
 	return 0;
 }
 
-bool hashmap_find(const struct hashmap *map, long key, long *value)
+bool hashmap_find(const struct hashmap *map, intptr_t key, intptr_t *value)
 {
 	struct hashmap_entry *entry;
 	size_t h;
@@ -217,8 +217,8 @@ bool hashmap_find(const struct hashmap *map, long key, long *value)
 	return true;
 }
 
-bool hashmap_delete(struct hashmap *map, long key,
-		    long *old_key, long *old_value)
+bool hashmap_delete(struct hashmap *map, intptr_t key,
+		    intptr_t *old_key, intptr_t *old_value)
 {
 	struct hashmap_entry **pprev, *entry;
 	size_t h;
